requirements.txt
# -*- coding: utf-8 -*-
# 工程预装模块
# 需要额外的python包，可直接在文件后面添加
# 请确保指定的包和版本号，可通过pip安装


django==2.0.6
pymysql==0.9.3
django-cors-headers==2.5.2
djangorestframework==3.8.2